Add model utils and moved isRedKing there
